Add SENTRY_DSN environment variable to production.py
This allows Sentry alerts to be sent to an alternate Sentry instance if one is set. This can be useful when someone wishes to maintain a fork. 
(Arguably the default should be set to null, disabling the feature all together, but this would likely be breaking and hence warrant further discussion.)